feat: prepare cart hook
