+++ src/components/hooks/use-shopping-cart/index.ts
@@ -1,0 +1,1 @@
+export { useShoppingCart } from './use-shopping-cart';
